move interface file to helper folder
